requirements.txt
pymupdf
shapely
langchain-community
langchain_openai
langchain_experimental
torch
IPython
pymupdf4llm
GeneralAgent
GeneralAgent
paddleocr
paddlepaddle